Change internal sequence counter of DuckDbAppender to 64-bit
The counter is allowed to wrap around so there was no real issue with 32-bit counters.

The counter does not need to be read or written atomically, so the
hypothetical issue of it being less efficient for 32-bit platforms is
irrelevant anyway.

+++ Mallard/Appender/DuckDbAppender.Slot.cs
@@ -45,7 +45,7 @@ public partial class DuckDbAppender
         /// are used.
         /// </para>
         /// </remarks>
-        private readonly uint _sequenceCounter;
+        private readonly ulong _sequenceCounter;
 
         private void CheckSequenceCounter()
         {
